"Route" class added and designed - "CountryMap" class designed
I designed a “Route” class (again as part of object-oriented approach) which will store the route between two specified cities and the travel time. It also includes a constructor method and getter methods for the two specified cities and time.

I also designed the “CountryMap” class which will store the cities and routes in separate arrays and the variables cityArrayCounter and routeArrayCounter are used to keep track of the number of cities and routes currently stored in each of the arrays.
The methods in the “CountryMap” class include:
* A constructor method to initialize the variables.
* A “cityExists” method to check if a city with the given name exists in the cities array. (I designed this method to use in the “addCity” method.)
* A “addCity” method to add a new city with the given name. (If it does not already exist and there is space in the array.)
* A “noCity” method to create a new City object with the name “No city”. (I designed this method to use in the “getCity” and “addRoute” methods.)
* A “getCity” method to get the name of a specific city in the cities array. If the city is not found in the array, it uses the “noCity” method to set the city with the name “No city”. (I designed this class to use in the “addRoute” method.)
* A “addRoute” method to add a new route between two specified cities with the given travel time. (If both cities exist and there is space in the array.)
* Getter methods for cities and routes.